Refractored
